pyproject.toml
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  .eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | migrations
  | scripts
)
'''

[tool.poetry]
name = "django-coding-challenge"
version = "0.1.0"
description = ""
authors = ["Tim Fromme <tfromme@cytracom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
argon2-cffi = "^23.1.0"
uvicorn = { extras = ["standard"],  version = "^0.29.0" }
redis = { version = "^5", extras = ["hiredis"] }
Django = "^4.2"
Jinja2 = "^3"
django-redis = "^5"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15"



[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
Werkzeug = "^3"
ipdb = "^0.13"
psycopg = "^3.1"
watchfiles = "^0.21.0"
mypy = "^1.9.0"
django-stubs = "^4.2.7"
pytest = "^8.1.1"
pytest-sugar = "^1.0.0"
djangorestframework-stubs = "^3.14.5"
flake8 = "^7.0.0"
flake8-isort = "^6.1.1"
black = "^24.3.0"
pylint-django = "^2.5.5"
factory-boy = "^3.3.0"
django-extensions = "^3.2.3"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
